chore: save and transform chats returned by login in ChatsToSellUnfolded
